Added MATLAB code snippet for root locus
The listings package's \lstlistlistings doesn't render correctly in the table
of contents, so a custom snippet environment was created with the same usage
and rendering style as figures. A typo in the example transfer function for
root locus was also fixed.

C++ syntax highlighting was also included, but is currently unused.