Improve state-space example scripts
The README sections now have more descriptive names and instructions on
running the examples are now easier to follow.

The examples are now interactive by default. With the --noninteractive
flag, examples output the plots to files, but don't display them to the
user. Pole-zero maps for examples that require Slycot to find the
transmission zeroes were disabled.