Make the build parallelizable
Instead of using svg2pdf.py, which operates on global state like the
current state of the directory tree, the scripts in the code directory now
convert their outputs to PDF directly using inkscape. The frccontrol
example execution was also moved into the build/frccontrol rule so it
would only be executed once when the folder needs to be created.

Now the -j flag can be passed to make with parallelized script executions
and latex compilations happening in the correct order.