Add defernumbers=true to biblatex arguments
This argument defers citation numbering to the location at which the
bibliography is generated, which produces a monotonically increasing
list of entry numbers throughout the bibliography subsections. This
makes finding the bibliography entry for a given citation easier.

Note that citation numbers are now no longer in the order they appear in
the main text, but that's okay because the more important property is
that the bibliography entries are in order so they're easier to scan.